requirements.txt
streamlit
python-dotenv
pypdf
langchain
langchain-text-splitters
langchain-openai
langchain-community
openai
faiss-cpu
